[v 3.1.11] - Changes architecture
- Move classes in differents files
- Move files in directories
- fix imports
